Se incluye permisos recurrente y mensaje de error para carpeta vacia

+++ admin/CargaMasivaSmartSearch_LS_01/CargaMasivaPDFxCargaMasiva.sh
@@ -6,6 +6,7 @@ sql_password="*ubuntuCarga123"
 sql_database="haptrack_docu"
 url_smart="/var/www/html/smartsearc/swdoc/HapAdmcom/docs/"
 
+chmod 777 -R PDFxCargaMasiva
 cd PDFxCargaMasiva
 #Obtiene información de cargue cliente carpeta y tipo de archivo del txt _ClienteCarpetaCarga.txt
 while read LINEA; do
@@ -22,11 +23,15 @@ done < ../_ClienteCarpetaCarga.info
 #obtiene los archivos a cargar
 for f in *.*
 do
-  echo $f
-  sql_insert=$sql_insert"INSERT INTO documentos (idcliente, fecha, nombre, adjunto, padre, tipo, tamano) values ('${PARTES[0]}', now(),'$f', '${PARTES[0]}"/"${PARTES[1]}"/"$f','${PARTES[1]}', 'application/${f#*.}','$(stat -c %s "$f")'); ";
-  mv "$f" $url_smart${PARTES[0]}"/"${PARTES[1]}"/$f";
+  if [ ! -f $f ];
+  then
+    echo "!!!!.....Carpeta PDFxCargaMasiva vacia. No hay archivos a cargar...!!!! Verifique!!"
+  else
+    sql_insert=$sql_insert"INSERT INTO documentos (idcliente, fecha, nombre, adjunto, padre, tipo, tamano) values ('${PARTES[0]}', now(),'$f', '${PARTES[0]}"/"${PARTES[1]}"/"$f','${PARTES[1]}', 'application/${f#*.}','$(stat -c %s "$f")'); ";
+    mv "$f" $url_smart${PARTES[0]}"/"${PARTES[1]}"/$f";
+    echo "..Procesando..."$f
+  fi
 done
-echo "..Procesando..."
 
 #Conectarse a la BD
 sql_args="-h $sql_host -u $slq_usuario -p$sql_password -D $sql_database -s -e"
